basic fishing is complete
