+++ ewcfg.py
@@ -79,7 +79,9 @@ poi_id_neomilwaukeestate = "neomilwaukeestate"
 poi_id_beachresort = "thebeachresort"
 poi_id_countryclub = "thecountryclub"
 poi_id_slimesea = "slimesea"
-poi_id_thepier = "thepier"
+poi_id_se_pier = "slimesendpier"
+poi_id_jp_pier = "jaywalkerplainpier"
+poi_id_lc_pier = "littlechernobylpier"
 
 # transports
 poi_id_ferry = "ferry"
@@ -262,7 +264,9 @@ channel_countryclub = "the-country-club"
 channel_rowdyroughhouse = "rowdy-roughhouse"
 channel_copkilltown = "cop-killtown"
 channel_slimesea = "slime-sea"
-channel_thepier = "the-pier"
+channel_se_pier = "slimes-end-pier"
+channel_jp_pier = "jaywalker-plain-pier"
+channel_lc_pier = "little-chernobyl-pier"
 
 channel_wt_port = "wreckington-port"
 channel_vc_port = "vagrants-corner-port"
@@ -1821,7 +1825,7 @@ food_list = [
     EwFood(
         id_food="water",
         alias=[
-            "h20",
+            "h2o",
         ],
         recover_hunger=0,
         price=0,
@@ -4383,21 +4387,62 @@ poi_list = [
         is_subzone=True,
         mother_district=poi_id_dreadford
     ),
-    EwPoi(  # The Pier
-        id_poi=poi_id_thepier,
+    EwPoi(  # Slime's End Pier
+        id_poi=poi_id_se_pier,
         alias=[
             "pier",
             "thepier",
             "the pier",
+            "sep",
+            "sepier",
+            "slimesendpier"
         ],
-        str_name="The Pier",
-        str_desc="A Shoddily built pile of wood planks and nails hangs off the edge of the peninsula.\n\nExits into Slime's End.",
+        str_name="Slime's End Pier",
+        str_desc="A shoddily built pile of wood planks and nails hangs off the edge of the peninsula.\n\nExits into Slime's End.",
         coord = (42, 16),
-        channel=channel_thepier,
-        role="The Pier",
+        channel=channel_se_pier,
+        role="Slime's End Pier",
         pvp=False,
         is_subzone=True,
         mother_district=poi_id_slimesend
+    ),
+    EwPoi(  # Jaywalker Plain Pier
+        id_poi=poi_id_jp_pier,
+        alias=[
+            "jpp",
+            "jppier",
+            "jp pier",
+            "jaywalkerplainpier",
+            "jaywalker plain pier"
+        ],
+        str_name="Jaywalker Plain Pier",
+        str_desc="A shoddily built pile of wood planks and nails hangs off the edge of the peninsula.\n\nExits into Jaywalker Plain.",
+        coord = (2, 18),
+        channel=channel_jp_pier,
+        role="Jaywalker Plain Pier",
+        pvp=False,
+        is_subzone=True,
+        mother_district=poi_id_jaywalkerplain
+    ),
+    EwPoi(  # Little Chernobyl Pier
+        id_poi=poi_id_lc_pier,
+        alias=[
+            "lcp",
+            "lcpier",
+            "lc pier",
+            "littlechernobylpier",
+            "lilchernpier",
+            "lilpier",
+            "lilp"
+        ],
+        str_name="Little Chernobyl Pier",
+        str_desc="A shoddily built pile of wood planks and nails hangs off the edge of the peninsula.\n\nExits into Little Chernobyl.",
+        coord = (24, 8),
+        channel=channel_lc_pier,
+        role="Little Chernobyl Pier",
+        pvp=False,
+        is_subzone=True,
+        mother_district=poi_id_littlechernobyl
     ),
     EwPoi(  # Slime Sea
         id_poi=poi_id_slimesea,
@@ -4753,7 +4798,7 @@ poi_list = [
         ],
         str_name="The Jaywalker Plain Subway Station",
         str_desc=str_green_subway_station_description + "\n\nExits into Jaywalker Plain.",
-        coord=(2, 19),
+        coord=(2, 20),
         channel=channel_jp_subway_station,
         role="Jaywalker Plain Subway Station",
         pvp=True,
@@ -7266,601 +7311,7 @@ for m in cosmetic_items_list:
     else:
         pass
 
-class EwFish:
-    # A unique name for the fish. This is used in the database and typed by
-    # users, so it should be one word, all lowercase letters.
-    id_fish = ""
 
-    # Name of fish.
-    str_name = ""
-
-    # Size of fish.
-    str_size = ""
-
-    # When it can be caught.
-    catch_time = None
-
-    # Flavor text upon catch.
-    flavor = ""
-
-    def __init__(
-            self,
-            id_fish="",
-            str_name="",
-            str_size="",
-            catch_time=None,
-            flavor=""
-    ):
-        self.id_fish = id_fish
-        self.str_name = str_name
-        self.str_size = str_size
-        self.catch_time = catch_time
-        self.flavor = flavor
-
-# All the fish, baby!
-fish_list = [
-    EwFish(
-        id_fish="neoneel",
-        str_name="Neon Eel",
-        str_size="a",
-        catch_time=None,
-        flavor="Its slippery body is bathed in a bright green glow."
-    ),
-    EwFish(
-        id_fish="fantaray",
-        str_name="Fanta Ray",
-        str_size="h",
-        catch_time=None,
-        flavor="Wait a minute, wasn't this the thing that killed that famous guy? Better be careful!"
-    ),
-    EwFish(
-        id_fish="thalamuscaranx",
-        str_name="Thalamus Caranx",
-        str_size="",
-        catch_time="n",
-        flavor="Finally, a worthy fish emerges."
-    ),
-    EwFish(
-        id_fish="fuckshark",
-        str_name="Fuck Shark",
-        str_size="",
-        catch_time=None,
-        flavor="You recall reading that this thing has the same nutritional value as SUPER WATER FUCK ENERGY."
-    ),
-    EwFish(
-        id_fish="sourfish",
-        str_name="Sourfish",
-        str_size="m",
-        catch_time=None,
-        flavor="It gives you an oddly cynical gaze."
-    ),
-    EwFish(
-        id_fish="snakeheadtrout",
-        str_name="Snakehead Trout",
-        str_size="a",
-        catch_time="n",
-        flavor="It has the body of a trout and the head of a snake. Heavy fuckin' metal."
-    ),
-    EwFish(
-        id_fish="gar",
-        str_name="Gar",
-        str_size="b",
-        catch_time=None,
-        flavor="You have the strange urge to wrestle this fish into submission. You almost resist it."
-    ),
-    EwFish(
-        id_fish="clownfish",
-        str_name="Clownfish",
-        str_size="m",
-        catch_time="d",
-        flavor="Its face kinda looks like a clown if you squint."
-    ),
-    EwFish(
-        id_fish="seasaint",
-        str_name="Seasaint",
-        str_size="m",
-        catch_time="n",
-        flavor="It has a beanie on."
-    ),
-    EwFish(
-        id_fish="holykrakerel",
-        str_name="Holy Krakerel",
-        str_size="",
-        catch_time="n",
-        flavor="It looks bovine-adjacent."
-    ),
-    EwFish(
-        id_fish="seajuggalo",
-        str_name="Sea Juggalo",
-        str_size="b",
-        catch_time="d",
-        flavor="This motherfucker definitely has some sick fuckin' musical taste."
-    ),
-    EwFish(
-        id_fish="plebefish",
-        str_name="Plebefish",
-        str_size="m",
-        catch_time=None,
-        flavor="God. This fucking retard. It just doesn't fucking GET it."
-    ),
-    EwFish(
-        id_fish="bufferfish",
-        str_name="Bufferfish",
-        str_size="",
-        catch_time=None,
-        flavor="This fish has the ability to lag out predators in order to get away."
-    ),
-    EwFish(
-        id_fish="slimesquid",
-        str_name="Slime Squid",
-        str_size="a",
-        catch_time=None,
-        flavor="It's just a green squid."
-    ),
-    EwFish(
-        id_fish="jellyturkeyfish",
-        str_name="Jelly Turkeyfish",
-        str_size="a",
-        catch_time=None,
-        flavor="You nearly prick your finger on one of the many of the venomous spines on its back."
-    ),
-    EwFish(
-        id_fish="iridescentsnapper",
-        str_name="Iridescent Snapper",
-        str_size="b",
-        catch_time=None,
-        flavor="Its scales change color if you shake it. Fun."
-    ),
-    EwFish(
-        id_fish="barredkatanajaw",
-        str_name="Barred Katanajaw",
-        str_size="b",
-        catch_time=None,
-        flavor="Its stripes make it look vaguely Japanese."
-    ),
-    EwFish(
-        id_fish="doublestuffedflounder",
-        str_name="Double-Stuffed Flounder",
-        str_size="c",
-        catch_time=None,
-        flavor="No one out-Flounders this fish."
-    ),
-    EwFish(
-        id_fish="seacolonel",
-        str_name="Sea Colonel",
-        str_size="c",
-        catch_time=None,
-        flavor="This fish definitely looks like its dropped out of high school."
-    ),
-    EwFish(
-        id_fish="marlinsupreme",
-        str_name="Marlin Supreme",
-        str_size="c",
-        catch_time=None,
-        flavor="Live mas."
-    ),
-    EwFish(
-        id_fish="relicanth",
-        str_name="Relicanth",
-        str_size="h",
-        catch_time="r",
-        flavor="It doesn't have teeth."
-    ),
-    EwFish(
-        id_fish="stunfisk",
-        str_name="h",
-        str_size="Stunfisk",
-        catch_time="r",
-        flavor="Its hide is so tough it can be stepped on by Connor without being injured."
-    ),
-    EwFish(
-        id_fish="bathyphysaheadshark",
-        str_name="Bathyphysahead Shark",
-        str_size="c",
-        catch_time=None,
-        flavor="This one looks fucking terrifying. I'm serious, search for 'bathyphysa' on Google."
-    ),
-    EwFish(
-        id_fish="anglershark",
-        str_name="Angler Shark",
-        str_size="h",
-        catch_time="n",
-        flavor="It has a little poudrin on its head."
-    ),
-    EwFish(
-        id_fish="bigtopoctopus",
-        str_name="Big Top Octopus",
-        str_size="h",
-        catch_time="d",
-        flavor="It kinda looks like a circus tent."
-    ),
-    EwFish(
-        id_fish="souroctopus",
-        str_name="Sour Octopus",
-        str_size="b",
-        catch_time=None,
-        flavor="It would rather be in a jar."
-    ),
-    EwFish(
-        id_fish="octohuss",
-        str_name="Octohuss",
-        str_size="c",
-        catch_time=None,
-        flavor="Don't let it near a horse. Or a drawing tablet."
-    ),
-    EwFish(
-        id_fish="jarocephalopod",
-        str_name="Jar O' Cephalopod",
-        str_size="a",
-        catch_time=None,
-        flavor="It looks content in there."
-    ),
-    EwFish(
-        id_fish="dab",
-        str_name="Dab",
-        str_size="a",
-        catch_time="n",
-        flavor="Pretty Killercore."
-    ),
-    EwFish(
-        id_fish="thrash",
-        str_name="Thrash",
-        str_size="a",
-        catch_time="d",
-        flavor="Pretty Rowdycore."
-    ),
-    EwFish(
-        id_fish="arsonfish",
-        str_name="Arsonfish",
-        str_size="a",
-        catch_time=None,
-        flavor="Its scales are so hot, you continuously toss the fish upwards to avoid getting burned."
-    ),
-    EwFish(
-        id_fish="cruna",
-        str_name="Cruna",
-        str_size="a",
-        catch_time=None,
-        flavor="It's just a green tuna fish."
-
-    ),
-    EwFish(
-        id_fish="modelopole",
-        str_name="Modelopole",
-        str_size="a",
-        catch_time=None,
-        flavor="UH-OH, IT'S MODELOPOLE TIME!"
-    ),
-    EwFish(
-        id_fish="universefrog",
-        str_name="Universe Frog",
-        str_size="c",
-        catch_time=None,
-        flavor="It's a huge fuckin' color-changing frog."
-    ),
-    EwFish(
-        id_fish="galaxyfrog",
-        str_name="Galaxy Frog",
-        str_size="h",
-        catch_time=None,
-        flavor="It's a big fuckin' color-changing frog."
-    ),
-    EwFish(
-        id_fish="solarfrog",
-        str_name="Solar Frog",
-        str_size="h",
-        catch_time="d",
-        flavor="Don't stare at it!"
-    ),
-    EwFish(
-        id_fish="lunarfrog",
-        str_name="Lunar Frog",
-        str_size="h",
-        catch_time="n",
-        flavor="It's said to control the waves of the Slime Sea."
-    ),
-    EwFish(
-        id_fish="killifish",
-        str_name="Killifish",
-        str_size="a",
-        catch_time="n",
-        flavor="Apparently there are 1270 different species of Killifish."
-    ),
-    EwFish(
-        id_fish="lee",
-        str_name="Lee",
-        str_size="",
-        catch_time=None,
-        flavor="Oh shit, it's Lee!"
-    ),
-    EwFish(
-        id_fish="palemunch",
-        str_name="Pale Munch",
-        str_size="a",
-        catch_time="d",
-        flavor="This fish looks like it needs some sleep."
-    ),
-    EwFish(
-        id_fish="moldfish",
-        str_name="Moldfish",
-        str_size="a",
-        catch_time=None,
-        flavor="It's said to have the memory capacity of 16 GB."
-    ),
-    EwFish(
-        id_fish="neonjuvie",
-        str_name="Neon Juvie",
-        str_size="a",
-        catch_time=None,
-        flavor="Pretty Juviecore."
-    ),
-    EwFish(
-        id_fish="greengill",
-        str_name="Greengill",
-        str_size="",
-        catch_time=None,
-        flavor="Its gills are green."
-    ),
-    EwFish(
-        id_fish="corpsecarp",
-        str_name="Corpse Carp",
-        str_size="a",
-        catch_time=None,
-        flavor="It smells like a rotting fish."
-    ),
-    EwFish(
-        id_fish="slimewatergoby",
-        str_name="Slimewater Goby",
-        str_size="m",
-        catch_time=None,
-        flavor="This little fucko hates fun."
-    ),
-    EwFish(
-        id_fish="nibblefish",
-        str_name="Nibblefish",
-        str_size="m",
-        catch_time=None,
-        flavor="It looks hungry."
-    ),
-    EwFish(
-        id_fish="piranhoid",
-        str_name="Piranhoid",
-        str_size="",
-        catch_time=None,
-        flavor="This fish is said to occasionally jump out of the water and bite unsuspecting slimeoids."
-    ),
-    EwFish(
-        id_fish="torrentfish",
-        str_name="Torrentfish",
-        str_size="b",
-        catch_time=None,
-        flavor="This fish looks like it doesn't pay for ANY of its anime."
-    ),
-    EwFish(
-        id_fish="barbeln8",
-        str_name="Barbel N8",
-        str_size="a",
-        catch_time="n",
-        flavor="It looks like it could run a shady corporation."
-    ),
-    EwFish(
-        id_fish="mace",
-        str_name="Mace",
-        str_size="",
-        catch_time=None,
-        flavor="These fish are called Mud Carps in Nu Hong Kong."
-    ),
-    EwFish(
-        id_fish="blacklimesalmon",
-        str_name="Black Lime Salmon",
-        str_size="h",
-        catch_time=None,
-        flavor="Kinda smells like Black Limes."
-    ),
-    EwFish(
-        id_fish="char",
-        str_name="Char",
-        str_size="",
-        catch_time=None,
-        flavor="These fish migrated south after the North Pole was nuked."
-    ),
-    EwFish(
-        id_fish="arijuana",
-        str_name="Arijuana",
-        str_size="",
-        catch_time=None,
-        flavor="These fish are banned from the USA."
-    ),
-    EwFish(
-        id_fish="thebassedgod",
-        str_name="The Bassed God",
-        str_size="c",
-        catch_time=None,
-        flavor="This is The Bassed God. He's gonna fuck your bitch."
-    ),
-    EwFish(
-        id_fish="flarp",
-        str_name="Flarp",
-        str_size="",
-        catch_time=None,
-        flavor="It's a carp thats really flexible."
-    ),
-    EwFish(
-        id_fish="clouttrout",
-        str_name="Clout Trout",
-        str_size="a",
-        catch_time=None,
-        flavor="This fish has the eyes of a winner."
-    ),
-    EwFish(
-        id_fish="slimekoi",
-        str_name="Slimekoi",
-        str_size="a",
-        catch_time=None,
-        flavor="Slimekoi is a level 3 slimeboi."
-    ),
-    EwFish(
-        id_fish="deadkoi",
-        str_name="Deadkoi",
-        str_size="a",
-        catch_time=None,
-        flavor="Deadkoi is a level 3 deadboi."
-    ),
-    EwFish(
-        id_fish="magicksdorado",
-        str_name="magicksDorado",
-        str_size="b",
-        catch_time="n",
-        flavor="No relation."
-    ),
-    EwFish(
-        id_fish="straubling",
-        str_name="Straubling",
-        str_size="b",
-        catch_time="d",
-        flavor="No relation."
-    ),
-    EwFish(
-        id_fish="croach",
-        str_name="Croach",
-        str_size="",
-        catch_time=None,
-        flavor="It's very uncommon in North America."
-    ),
-    EwFish(
-        id_fish="slimesmelt",
-        str_name="Slime Smelt",
-        str_size="a",
-        catch_time=None,
-        flavor="It could sure use a bath."
-    ),
-    EwFish(
-        id_fish="neomilwaukianmittencrab",
-        str_name="Neo-Milwaukian Mitten Crab",
-        str_size="b",
-        catch_time=None,
-        flavor="Known for their furry claws, Mitten Crabs were considered an invasive species, but eventually people stopped caring about that because they had bigger fish to fry (metaphorically, of course)."
-    ),
-    EwFish(
-        id_fish="yellowslash",
-        str_name="Yellow Slash",
-        str_size="",
-        catch_time=None,
-        flavor="This fish is the successor to Classic Milwaukee's Yellow Perch."
-    ),
-    EwFish(
-        id_fish="sweetfish",
-        str_name="sweet fish",
-        str_size="m",
-        catch_time="d",
-        flavor="Also known as Gillanaks."
-    ),
-    EwFish(
-        id_fish="hardboiledturtle",
-        str_name="Hard Boiled Turtle",
-        str_size="a",
-        catch_time=None,
-        flavor="This radical dude doesn't take shit from anyone."
-    ),
-    EwFish(
-        id_fish="oozesalmon",
-        str_name="Ooze Salmon",
-        str_size="a",
-        catch_time=None,
-        flavor="You wonder how good it would taste on a bagel."
-    ),
-    EwFish(
-        id_fish="toxicpike",
-        str_name="Toxic Pike",
-        str_size="a",
-        catch_time=None,
-        flavor="Don't let it bite you."
-    ),
-    EwFish(
-        id_fish="kingpincrab",
-        str_name="Kingpin Crab",
-        str_size="h",
-        catch_time=None,
-        flavor="It reminds you of your last meal at Red Mobster."
-    ),
-    EwFish(
-        id_fish="regiarapaima",
-        str_name="Regiarapaima",
-        str_size="c",
-        catch_time=None,
-        flavor="Regigas sends its regards."
-    ),
-    EwFish(
-        id_fish="kinkfish",
-        str_name="Kinkfish",
-        str_size="h",
-        catch_time=None,
-        flavor="This fish looks like it's down to get wacky."
-    ),
-    EwFish(
-        id_fish="nuclearbream",
-        str_name="Nuclear Bream",
-        str_size="",
-        catch_time=None,
-        flavor="Not to be confused with BREEAM, although this fish looks like its in the mood for assessing shit."
-    ),
-    EwFish(
-        id_fish="killercod",
-        str_name="Killer Cod",
-        str_size="a",
-        catch_time="n",
-        flavor="Quite Killercore."
-    ),
-    EwFish(
-        id_fish="pinksnapper",
-        str_name="Pink Snapper",
-        str_size="a",
-        catch_time="d",
-        flavor="Quite Rowdycore."
-    ),
-    EwFish(
-        id_fish="angerfish",
-        str_name="Angerfish",
-        str_size="h",
-        catch_time=None,
-        flavor="It doesn't look very happy to be here."
-    ),
-    EwFish(
-        id_fish="flopfish",
-        str_name="Flop Fish",
-        str_size="b",
-        catch_time=None,
-        flavor="It's floppin'."
-    ),
-    EwFish(
-        id_fish="cardboardcrab",
-        str_name="Cardboard Crab",
-        str_size="b",
-        catch_time=None,
-        flavor="It originated when Shigeru Miyamoto decided to splice crab DNA with Nintendo Labo Piano."
-    ),
-    EwFish(
-        id_fish="easysardines",
-        str_name="Easy Sardines",
-        str_size="m",
-        catch_time=None,
-        flavor="In terms of difficulty, this little bitch looks real low on the rungs."
-    ),
-    EwFish(
-        id_fish="largebonedlionfish",
-        str_name="Large-Boned Lionfish",
-        str_size="a",
-        catch_time=None,
-        flavor="It's not fat."
-    ),
-    EwFish(
-        id_fish="paradoxcrocodile",
-        str_name="Paradox Crocodile",
-        str_size="c",
-        catch_time=None,
-        flavor="It has no arms and a blue bandana."
-    )
-]
 
 # Dict of all help responses linked to their associated topics
 help_responses = {
