+++ ewcmd.py
