+++ ewmap.py
@@ -13,6 +13,7 @@ from ew import EwUser
 from ewdistrict import EwDistrict
 from ewtransport import EwTransport
 from ewslimeoid import EwSlimeoid
+from ewfishy import fishers
 
 move_counter = 0
 
@@ -183,8 +184,8 @@ map_world = [
 	[ -1, -1, -1, -1, -1,  0,  0,  0, 30, -2, 30,  0,  0,  0, -1, -1, -1, -3, -1, -1, -1, 30, -1, -1, -1,  0, -1, -1, 30, -1, -1, -1, -1,  0,  0, 30, -2, 30,  0, 30, -3, 20, -2, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1, -1,  0, -1, -1, -1, 20, -1, -1, -1,  0,  0,  0, -1, -3, -1,  0,  0,  0,  0, -1, -1,  0, -1, -1,  0, -1, -1, -1, -1, -1, -1, -1, 30, -1, -1, -1, -3, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1, -1,  0, -1, -1, -1, -2, -1, -1, -1, -1, -1,  0, 30, -2, 30,  0, -1, -1,  0, -1, -1, 30, -1, -1,  0, -1, -1, -1, -1, -1, -1, -1,  0, -1, -1, 30, -2, 20, -2, -1, -1, -1, -1, -1, -1, -1 ],
-	[ -1, -1, -1, -1, -1,  0, -1, -1, -1, -1, -1, -1, -1,  0,  0,  0, -1, 30, -1, -1, -1, -1,  0, -1, -1, -2, 30,  0,  0,  0, -1, -1, -1, -1, -1, -1,  0, -1, -1,  0, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
-	[ -1, -1, -1, -1, -1, 30, -1,  0,  0,  0,  0,  0, -1, 30, -1, -1, -1,  0, -1, -1, -1, -1,  0, -1, -1, -1, -1, -1, -1,  0, -1, -1, -1, -1, -1, -1,  0, -1, -1, 30, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
+	[ -1, -1, -1, -1, -1,  0, -1, -1, -1, -1, -1, -1, -1,  0,  0,  0, -1, 30, -1, -1, -1, -1,  0, -1, 20, -2, 30,  0,  0,  0, -1, -1, -1, -1, -1, -1,  0, -1, -1,  0, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
+	[ -1, -1, -1, -1, -1, 30, -1,  0,  0,  0,  0,  0, -1, 30, -1, -1, -1,  0, -1, -1, -1, -1,  0, -1, -2, -1, -1, -1, -1,  0, -1, -1, -1, -1, -1, -1,  0, -1, -1, 30, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1, 30, -2, 30,  0, -1, -1, -1,  0, 30, -2, 20, -2, -1,  0, -1, -1, -1, -1,  0,  0, -1, -2, -1, -1, -1,  0,  0, 30, -2, 30,  0,  0,  0,  0, -1, -3, 20, -2, -1, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1,  0, -1, -1,  0, -1, -1, -2, -1, -1, -3, -1, -1, -1,  0, -1, -1, -1, -1, -1, 30, -1, 20, -1, -1, -1, -1, -1, -1, 30, -1, -1, -1, -1, 30, -1, -3, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1,  0, -1, -1,  0, -1, -1, 20, -1, -1, -3, 20, -2, -1,  0, -1, -1, -1, -2, 20, -2, -3, -3, 30,  0,  0,  0,  0,  0,  0, -1,  0,  0, 30, -2, -3, -3, 20, -2, -1, -1, -1, -1, -1, -1, -1, -1 ],
@@ -194,9 +195,9 @@ map_world = [
 	[ -1, -1, -1, -1, 30, -1, -1, -1, 30, -1, -1, -1, -1, -1, -1, -1, -1, -2, -1, -1, -1, -1, -1, 30, -1, 30, -1, -1, -2, -1, -2, -1,  0, -1, -1, -1, -1,  0, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1,  0, 30, -3, -3, -2, 30,  0,  0,  0,  0,  0, -1, -1, -1, -1, -1, -1, -2, 20, -2, -3, -3, 20, -2, -1, -1, -1, -1,  0, -1, -1, -1, -1,  0,  0, 30, -2, 20, -2, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1,  0, -1, 20, -1, 30, -1,  0, -1, -1, -1,  0, -1, -2, -1, -2, -1, -1, -1, -1, -3, -1, -3, -1, -1, -1, -1, -1, -1, 30, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
-	[ -1, -1, -1, -1, 30, -1, -2, -1,  0, -1, 30, -1, -1, -1,  0, -1, 20, -1, 20, -1,  0,  0, 30, -3, -1, -3, 20, -2, -1, -1, -2, 20, -2, 20, -2, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
-	[ -1, -1, -2, 20, -2, -3, -1,  0,  0, 30, -2, 20, -2, -1,  0, 30, -2, -3, -3, 30,  0, -1, -1, 20, -1, 30, -1, -1, -1, -1, -1, -1, 20, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
-	[ -1, -1, -1, -1, -1, -3, 30,  0, -1, -1, -3, -1, -1, -1, -1, -1, 30, -1, -1, -1,  0, -1, -1, -2, -1, 30, -1, -1, -1, -1, -1, -1, -2, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
+	[ -1, -1, -2, -1, 30, -1, -2, -1,  0, -1, 30, -1, -1, -1,  0, -1, 20, -1, 20, -1,  0,  0, 30, -3, -1, -3, 20, -2, -1, -1, -2, 20, -2, 20, -2, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
+	[ -1, -1,  0, 20, -2, -3, -1,  0,  0, 30, -2, 20, -2, -1,  0, 30, -2, -3, -3, 30,  0, -1, -1, 20, -1, 30, -1, -1, -1, -1, -1, -1, 20, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
+	[ -1, -1, -2, -1, -1, -3, 30,  0, -1, -1, -3, -1, -1, -1, -1, -1, 30, -1, -1, -1,  0, -1, -1, -2, -1, 30, -1, -1, -1, -1, -1, -1, -2, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1, -1, 30, -1, -1,  0, 30, -3, 30, 30, -1, -1, -1,  0, -1, -1,  0,  0, -1, -1, -1, -1, -2, -3, -3, 20, -2, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -1, -1, -1, -1,  0, -1, -1, 30, -1, 20, -1, -2, 30,  0,  0,  0, -1, -1, 30, -1,  0,  0,  0,  0, 30, -1, 30, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
 	[ -1, -2, 20, -2, 30,  0,  0, 30, -2, -1, -2, -1, -3, -1,  0, -1, -1,  0, 30, -2, 30,  0, -1, -1, -1, -1, -1, 30, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 ],
@@ -579,6 +580,12 @@ async def move(cmd):
 		) if minutes > 0 else (" It's {} seconds away.".format(seconds) if seconds > 30 else ""))
 	)))
 
+	if cmd.message.author.id in fishers.keys():
+		fishers[cmd.message.author.id].fishing = False
+		fishers[cmd.message.author.id].bite = False
+		fishers[cmd.message.author.id].current_fish = ""
+		fishers[cmd.message.author.id].pier = ""
+
 	life_state = user_data.life_state
 	faction = user_data.faction
 
@@ -698,6 +705,11 @@ async def move(cmd):
 					user_data.persist()
 
 					await ewrolemgr.updateRoles(client = cmd.client, member = cmd.message.author)
+					if cmd.message.author.id in fishers.keys():
+						fishers[cmd.message.author.id].fishing = False
+						fishers[cmd.message.author.id].bite = False
+						fishers[cmd.message.author.id].current_fish = ""
+						fishers[cmd.message.author.id].pier = ""
 
 					try:
 						await cmd.client.delete_message(msg_walk_start)
