+++ ewquadrants.py
