+++ ewslimeoid.py
