+++ ewutils.py
